refactor: remove for now dependency on ensmallen
The ensmallen library has an optimizer that we might use to
find optimal values for the parameters of the system (i.e., alpha,
beta). But for now, I would prefer to try not to use it, as it
leads to additional dependencies that might be hard to install without
conan.